+++ src/aim/cftemplates/s3.py
@@ -113,12 +113,14 @@ Resources:
             Principal:
 {0[principal]:s}
             Resource: {0[resource_list]:s}
+            {0[condition]:s}
 """
         s3_policy_statement_table = {
             'action_list': "",
             'effect': "",
             'principal': "",
-            'resource_list': ""
+            'resource_list': "",
+            'condition': "",
             }
 
         s3_bucket_outputs_fmt = """
@@ -178,13 +180,39 @@ Resources:
                 # Action
                 for action in policy_statement.action:
                     s3_policy_statement_table['action_list'] += cf_list_item.format(action)
-                # Effict
+                # Effect
                 s3_policy_statement_table['effect'] = policy_statement.effect
-                # Principal
-                if policy_statement.aws != None and len(policy_statement.aws) > 0:
+
+                # Principal - principal field
+                if policy_statement.principal != None and len(policy_statement.principal) > 0:
+                    for key, value in policy_statement.principal.items():
+                        s3_policy_statement_table['principal'] = "              {}:".format(key)
+                        # can be a string or list
+                        if type(value) == type(list()):
+                            for item in value:
+                                s3_policy_statement_table['principal'] += cf_principal_list_item.format(item)
+                        else:
+                            s3_policy_statement_table['principal'] += cf_principal_list_item.format(value)
+
+                # Principal - aws field
+                elif policy_statement.aws != None and len(policy_statement.aws) > 0:
                     s3_policy_statement_table['principal'] = "              AWS:"
                     for principal in policy_statement.aws:
                         s3_policy_statement_table['principal'] += cf_principal_list_item.format(principal)
+
+                if policy_statement.condition != {}:
+                    padding = "              "
+                    condition = "Condition:\n"
+                    for key, value in policy_statement.condition.items():
+                        condition += padding + "  {}:\n".format(key)
+                        for sub_key, sub_value in value.items():
+                            if type(sub_value) == type(list()):
+                                condition += padding + "    '{}':\n".format(sub_key)
+                                for item in sub_value:
+                                    condition += padding + "      - '{}'\n".format(item)
+                            else:
+                                condition += padding + "    '{}': '{}'\n".format(sub_key, sub_value)
+                    s3_policy_statement_table['condition'] = condition
 
                 # Resource
                 bucket_arn = s3_ctl.get_bucket_arn(self.s3_context_id)